Refactor phonebook and add `PhonebookEntry.js`
I was having trouble finding generated list items' keys in the browser debugger
and eventually noticed that this info is available using the React dev tools'
element tree view, but only when the generated elements are React components.
The `<li>` elements this code was previously generating aren't components, and
therefore aren't available in the tree, and therefore the keys can't be inspected.
This new component now shows in the tree and displays the keys as expected